rename non-React files from .tsx to .ts
